Add send_message endpoints to telegram and VK api

+++ crates/domain_vk_bot/src/models.rs
@@ -5,8 +5,8 @@ use serde::Deserialize;
 pub struct VkCallbackRequest {
     #[serde(default)]
     pub r#type: VkCallbackType,
-    pub group_id: i32,
-    pub event_id: i32,
+    pub group_id: i64,
+    pub event_id: i64,
     pub secret: Option<String>,
     pub object: Option<NewMessageObject>,
 }
